# specs/001-backend-api-foundation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NALLO Backend API
  description: REST API for NALLO document management system
  version: 1.0.0
  contact:
    name: NALLO API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.nallo.example.com/api/v1
    description: Production server
...
paths:
...
  /concepts:
    post:
      tags:
        - Concepts
      summary: Create concept
      description: Create a new glossary concept
      operationId: createConcept
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConceptRequest'
      responses:
        '201':
          description: Concept created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConceptResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /concepts/{id}:
    get:
      tags:
        - Concepts
      summary: Get concept
      description: Retrieve concept by ID
      operationId: getConcept
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Concept retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConceptResponse'
        '404':
          description: Concept not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Concepts
      summary: Update concept
      description: Update existing concept
      operationId: updateConcept
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConceptRequest'
      responses:
        '200':
          description: Concept updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConceptResponse'
        '404':
          description: Concept not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /concepts/{id}/documents:
    get:
      tags:
        - Concepts
      summary: Get documents using concept
      description: Retrieve all documents that use this concept (impact analysis)
      operationId: getConceptDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of documents using concept
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '404':
          description: Concept not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [api, general, tutorial]
        status:
          type: string
          enum: [draft, in_review, done, publish]
        title:
          type: string
        lang:
          type: string
        content:
          type: string
        summary:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DocumentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    CreateConceptRequest:
      type: object
      required:
        - term
        - description
        - lang
      properties:
        term:
          type: string
        description:
          type: string
        category:
          type: string
        lang:
          type: string
          pattern: '^[a-z]{2}$'
    UpdateConceptRequest:
      type: object
      properties:
        term:
          type: string
        description:
          type: string
        category:
          type: string
    ConceptResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        term:
          type: string
        description:
          type: string
        category:
          type: string
        lang:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object